code cleanup
-moved hooks to proper file
-added on release cleanups
-changed target dll name
